Local branch amd-gfx 63e627b07a2b Merged main:e9b33d085da0 into amd-gfx:6510fa067355
Remote branch main f756061387c8 [Clang] Add missing words to ReleaseNotes.rst

+++ clang/lib/AST/Interp/IntegralAP.h
@@ -66,7 +66,7 @@ public:
 
   template <typename T> static IntegralAP from(T Value, unsigned NumBits = 0) {
     assert(NumBits > 0);
-    APSInt Copy = APSInt(APInt(NumBits, Value, Signed), !Signed);
+    APSInt Copy = APSInt(APInt(NumBits, static_cast<int64_t>(Value), Signed), !Signed);
 
     return IntegralAP<Signed>(Copy);
   }
